Implemented Monserrat Font and Hover Functionality
I got the correct font to show up on the webpage and implemented the hover functionality so that text shows up when the mouse is over certain board members' images. I'm waiting to hear back about whether any of the other board members have hover text to be implemented.
I was having trouble getting the hover function to work for the boardMembers class, so the three board members with hover are in a separate class called container. The board member's photos are not in the correct positions yet, but otherwise the page is looking nearly complete.